+++ opendc-common/src/main/kotlin/org/opendc/common/util/Pacer.kt
@@ -20,7 +20,7 @@
  * SOFTWARE.
  */
 
-package org.opendc.utils
+package org.opendc.common.util
 
 import kotlinx.coroutines.*
 import java.lang.Runnable
