+++ opendc-common/src/main/kotlin/org/opendc/common/util/TimerScheduler.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 AtLarge Research
+ * Copyright (c) 2022 AtLarge Research
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -20,7 +20,7 @@
  * SOFTWARE.
  */
 
-package org.opendc.utils
+package org.opendc.common.util
 
 import kotlinx.coroutines.*
 import kotlinx.coroutines.channels.Channel
