+++ opendc-compute/opendc-compute-service/build.gradle.kts
@@ -33,7 +33,7 @@ dependencies {
     api(platform(projects.opendcPlatform))
     api(projects.opendcCompute.opendcComputeApi)
     api(projects.opendcTelemetry.opendcTelemetryApi)
-    implementation(projects.opendcUtils)
+    implementation(projects.opendcCommon)
     implementation(libs.kotlin.logging)
     implementation(libs.opentelemetry.semconv)
