+++ opendc-compute/opendc-compute-simulator/build.gradle.kts
@@ -34,7 +34,7 @@ dependencies {
     api(projects.opendcCompute.opendcComputeService)
     api(projects.opendcSimulator.opendcSimulatorCompute)
     api(libs.commons.math3)
-    implementation(projects.opendcUtils)
+    implementation(projects.opendcCommon)
     implementation(libs.opentelemetry.semconv)
     implementation(libs.kotlin.logging)
