+++ opendc-faas/opendc-faas-service/build.gradle.kts
@@ -33,7 +33,7 @@ dependencies {
     api(platform(projects.opendcPlatform))
     api(projects.opendcFaas.opendcFaasApi)
     api(projects.opendcTelemetry.opendcTelemetryApi)
-    implementation(projects.opendcUtils)
+    implementation(projects.opendcCommon)
     implementation(libs.kotlin.logging)
     implementation(libs.opentelemetry.semconv)
